requirements.txt
astroid==2.9.3
autopep8==1.6.0
colorama==0.4.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
Pillow==9.0.1
platformdirs==2.5.1
pycodestyle==2.8.0
pylint==2.12.2
toml==0.10.2
typing_extensions==4.1.1
wrapt==1.13.3
